Add tracing feature
The tracing feature enables collection of scheduling events.
This data can be used by tracing tools (e.g. babeltrace2) running on a host
to visualize the kernel's inner workings and the behavior of its subsystems
(e.g. when a task is switched in or out).

Tracing events are stored in RAM in binary format and can later be
converted to human readable format using trace_convert.py script or
babeltrace2 with the bt_plugin_openbsw.py plugin.

Signed-off-by: Marcin Wierzbicki <marcin.wierzbicki@accenture.com>